theoraenc: do not use %zu, it is C99
Cast the variables instead and fallback to %u

+++ ext/theora/gsttheoraenc.c
@@ -851,8 +851,8 @@ theora_enc_buffer_from_header_packet (GstTheoraEnc * enc, ogg_packet * packet)
   GST_BUFFER_TIMESTAMP (outbuf) = GST_CLOCK_TIME_NONE;
   GST_BUFFER_DURATION (outbuf) = GST_CLOCK_TIME_NONE;
 
-  GST_DEBUG ("created header packet buffer, %zu bytes",
-      gst_buffer_get_size (outbuf));
+  GST_DEBUG ("created header packet buffer, %u bytes",
+      (guint) gst_buffer_get_size (outbuf));
   return outbuf;
 }
